requirements.txt
Django==4.2.23
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
gunicorn==23.0.0
requests==2.32.4
python-dateutil==2.9.0.post0
pytz==2025.2
